fix(ARC-432): fix prophecy duplicate card validation and split deity spec file (ARC-432)
