+++ apps/be/src/games/engines/critical/critical-deity.utils.spec.ts
@@ -4,8 +4,6 @@ import {
   CriticalPlayerState,
 } from '../../critical/critical.state';
 import {
-  executeResurrection,
-  executeJudgment,
   executeProphecy,
   executeCommitProphecy,
 } from './critical-deity.utils';
@@ -20,328 +18,36 @@ function makePlayer(
   return { playerId, hand, alive, stash: [], markedCards: [] };
 }
 
-function makeState(overrides: Partial<CriticalState> = {}): CriticalState {
-  return {
-    deck: [],
-    discardPile: [],
-    playerOrder: ['playerA', 'playerB'],
-    currentTurnIndex: 0,
-    pendingDraws: 1,
-    playDirection: 1,
-    expansions: ['deity'],
-    allowActionCardCombos: false,
-    pendingDefuse: null,
-    pendingFavor: null,
-    pendingAlter: null,
-    pendingAction: null,
-    eliminatedPlayers: [],
-    players: [
-      makePlayer('playerA', ['resurrection']),
-      makePlayer('playerB', [], false),
-    ],
-    logs: [],
-    ...overrides,
-  };
-}
-
 function makeHelpers(): EngineHelpers {
   return {
     addLog: (state: CriticalState, entry: GameLogEntry) => {
       state.logs.push(entry as CriticalState['logs'][number]);
     },
-    createLogEntry: (_type: string, message: string): GameLogEntry => ({
+    createLogEntry: (
+      _type: string,
+      message: string,
+      options?: Record<string, unknown>,
+    ): GameLogEntry => ({
       id: 'test-id',
       type: 'action',
       message,
       createdAt: new Date().toISOString(),
+      ...(options ?? {}),
     }),
-    advanceTurn: () => {},
+    advanceTurn: (state: CriticalState) => {
+      state.currentTurnIndex =
+        (state.currentTurnIndex + state.playDirection + state.playerOrder.length) %
+        state.playerOrder.length;
+      state.pendingDraws = 1;
+    },
     shuffleArray: () => {},
     findPlayer: (state: CriticalState, playerId: string) =>
       state.players.find((p) => p.playerId === playerId),
   };
 }
 
-describe('executeResurrection', () => {
-  it('revives the most recently eliminated player', () => {
-    const state = makeState({
-      eliminatedPlayers: ['playerB'],
-      deck: [
-        'strike',
-        'evade',
-        'reorder',
-        'strike',
-        'evade',
-        'reorder',
-        'insight',
-        'trade',
-        'cancel',
-        'insight',
-      ],
-    });
-    const helpers = makeHelpers();
-
-    const result = executeResurrection(state, 'playerA', helpers);
-
-    expect(result.success).toBe(true);
-    const revived = state.players.find((p) => p.playerId === 'playerB');
-    expect(revived?.alive).toBe(true);
-  });
-
-  it('removes the revived player from eliminatedPlayers', () => {
-    const state = makeState({
-      eliminatedPlayers: ['playerB'],
-      deck: [
-        'strike',
-        'evade',
-        'reorder',
-        'a',
-        'b',
-        'c',
-        'd',
-        'e',
-        'f',
-        'g',
-      ] as CriticalCard[],
-    });
-    const helpers = makeHelpers();
-
-    executeResurrection(state, 'playerA', helpers);
-
-    expect(state.eliminatedPlayers).not.toContain('playerB');
-    expect(state.eliminatedPlayers.length).toBe(0);
-  });
-
-  it('gives 3 cards from the bottom of the deck to the revived player', () => {
-    const deck: CriticalCard[] = [
-      'strike',
-      'evade',
-      'reorder',
-      'strike',
-      'evade',
-      'reorder',
-      'insight',
-      'cancel',
-      'trade',
-      'insight', // last 3: cancel, trade, insight
-    ];
-    const state = makeState({
-      eliminatedPlayers: ['playerB'],
-      deck: [...deck],
-    });
-    const helpers = makeHelpers();
-
-    executeResurrection(state, 'playerA', helpers);
-
-    const revived = state.players.find((p) => p.playerId === 'playerB');
-    expect(revived?.hand).toHaveLength(3);
-    expect(revived?.hand).toEqual(['cancel', 'trade', 'insight']);
-    expect(state.deck).toHaveLength(7);
-  });
-
-  it('removes exactly 3 cards from the deck', () => {
-    const deck: CriticalCard[] = Array(10).fill('strike') as CriticalCard[];
-    const state = makeState({
-      eliminatedPlayers: ['playerB'],
-      deck: [...deck],
-    });
-    const helpers = makeHelpers();
-
-    executeResurrection(state, 'playerA', helpers);
-
-    expect(state.deck).toHaveLength(7);
-  });
-
-  it('fails if eliminatedPlayers is empty', () => {
-    const state = makeState({
-      eliminatedPlayers: [],
-      deck: ['strike', 'evade', 'reorder'] as CriticalCard[],
-    });
-    const helpers = makeHelpers();
-
-    const result = executeResurrection(state, 'playerA', helpers);
-
-    expect(result.success).toBe(false);
-    expect(result.error).toBeDefined();
-  });
-
-  it('sets pendingAction correctly after resurrection', () => {
-    const state = makeState({
-      eliminatedPlayers: ['playerB'],
-      deck: ['strike', 'evade', 'reorder', 'a', 'b', 'c'] as CriticalCard[],
-    });
-    const helpers = makeHelpers();
-
-    executeResurrection(state, 'playerA', helpers);
-
-    expect(state.pendingAction).not.toBeNull();
-    expect(state.pendingAction?.type).toBe('resurrection');
-    expect(state.pendingAction?.playerId).toBe('playerA');
-    const payload = state.pendingAction?.payload as { targetPlayerId: string };
-    expect(payload.targetPlayerId).toBe('playerB');
-    expect(state.pendingAction?.nopeCount).toBe(0);
-  });
-
-  it('revives only the last eliminated player when multiple are eliminated', () => {
-    const state = makeState({
-      players: [
-        makePlayer('playerA', ['resurrection']),
-        makePlayer('playerB', [], false),
-        makePlayer('playerC', [], false),
-      ],
-      playerOrder: ['playerA', 'playerB', 'playerC'],
-      eliminatedPlayers: ['playerB', 'playerC'],
-      deck: Array(10).fill('strike') as CriticalCard[],
-    });
-    const helpers = makeHelpers();
-
-    executeResurrection(state, 'playerA', helpers);
-
-    const playerB = state.players.find((p) => p.playerId === 'playerB');
-    const playerC = state.players.find((p) => p.playerId === 'playerC');
-    expect(playerC?.alive).toBe(true);
-    expect(playerB?.alive).toBe(false);
-    expect(state.eliminatedPlayers).toEqual(['playerB']);
-  });
-
-  it('fails if the player executing resurrection does not exist', () => {
-    const state = makeState({
-      eliminatedPlayers: ['playerB'],
-      deck: Array(5).fill('strike') as CriticalCard[],
-    });
-    const helpers = makeHelpers();
-
-    const result = executeResurrection(state, 'nonExistentPlayer', helpers);
-
-    expect(result.success).toBe(false);
-  });
-});
-
-describe('executeJudgment', () => {
-  function makeJudgmentState(overrides: Partial<CriticalState> = {}): CriticalState {
-    return {
-      deck: [],
-      discardPile: [],
-      playerOrder: ['playerA', 'playerB', 'playerC'],
-      currentTurnIndex: 0,
-      pendingDraws: 1,
-      playDirection: 1,
-      expansions: ['deity'],
-      allowActionCardCombos: false,
-      pendingDefuse: null,
-      pendingFavor: null,
-      pendingAlter: null,
-      pendingAction: null,
-      eliminatedPlayers: [],
-      players: [
-        makePlayer('playerA', ['judgment']),
-        makePlayer('playerB', ['strike', 'evade', 'insight', 'trade', 'reorder']),
-        makePlayer('playerC', ['strike', 'evade', 'insight', 'trade']),
-      ],
-      logs: [],
-      ...overrides,
-    };
-  }
-
-  function makeAdvancingHelpers(): EngineHelpers {
-    return {
-      addLog: (state: CriticalState, entry: GameLogEntry) => {
-        state.logs.push(entry as CriticalState['logs'][number]);
-      },
-      createLogEntry: (_type: string, message: string): GameLogEntry => ({
-        id: 'test-id',
-        type: 'action',
-        message,
-        createdAt: new Date().toISOString(),
-      }),
-      advanceTurn: (state: CriticalState) => {
-        state.currentTurnIndex =
-          (state.currentTurnIndex + state.playDirection + state.playerOrder.length) %
-          state.playerOrder.length;
-        state.pendingDraws = 1;
-      },
-      shuffleArray: () => {},
-      findPlayer: (state: CriticalState, playerId: string) =>
-        state.players.find((p) => p.playerId === playerId),
-    };
-  }
-
-  it('sets pendingJudgment on all other alive players', () => {
-    const state = makeJudgmentState();
-    const helpers = makeAdvancingHelpers();
-
-    const result = executeJudgment(state, 'playerA', helpers);
-
-    expect(result.success).toBe(true);
-    const playerB = state.players.find((p) => p.playerId === 'playerB') as CriticalPlayerState;
-    const playerC = state.players.find((p) => p.playerId === 'playerC') as CriticalPlayerState;
-    expect(playerB.pendingJudgment).toBe(true);
-    expect(playerC.pendingJudgment).toBe(true);
-  });
-
-  it('does NOT set pendingJudgment on the player who played judgment', () => {
-    const state = makeJudgmentState();
-    const helpers = makeAdvancingHelpers();
-
-    executeJudgment(state, 'playerA', helpers);
-
-    const playerA = state.players.find((p) => p.playerId === 'playerA') as CriticalPlayerState;
-    expect(playerA.pendingJudgment).toBeUndefined();
-  });
-
-  it('skips dead players when setting pendingJudgment', () => {
-    const state = makeJudgmentState({
-      players: [
-        makePlayer('playerA', ['judgment']),
-        makePlayer('playerB', ['strike', 'evade', 'insight'], false),
-        makePlayer('playerC', ['strike', 'evade', 'insight', 'trade']),
-      ],
-    });
-    const helpers = makeAdvancingHelpers();
-
-    executeJudgment(state, 'playerA', helpers);
-
-    const playerB = state.players.find((p) => p.playerId === 'playerB') as CriticalPlayerState;
-    const playerC = state.players.find((p) => p.playerId === 'playerC') as CriticalPlayerState;
-    expect(playerB.pendingJudgment).toBeUndefined();
-    expect(playerC.pendingJudgment).toBe(true);
-  });
-
-  it('advances turn after playing judgment', () => {
-    const state = makeJudgmentState();
-    const helpers = makeAdvancingHelpers();
-
-    executeJudgment(state, 'playerA', helpers);
-
-    // Turn should have advanced from playerA (index 0) to playerB (index 1)
-    expect(state.currentTurnIndex).toBe(1);
-  });
-
-  it('sets pendingAction with type judgment', () => {
-    const state = makeJudgmentState();
-    const helpers = makeAdvancingHelpers();
-
-    executeJudgment(state, 'playerA', helpers);
-
-    expect(state.pendingAction).not.toBeNull();
-    expect(state.pendingAction?.type).toBe('judgment');
-    expect(state.pendingAction?.playerId).toBe('playerA');
-    expect(state.pendingAction?.nopeCount).toBe(0);
-  });
-
-  it('returns failure when player is not found', () => {
-    const state = makeJudgmentState();
-    const helpers = makeAdvancingHelpers();
-
-    const result = executeJudgment(state, 'nonExistent', helpers);
-
-    expect(result.success).toBe(false);
-    expect(result.error).toBeDefined();
-  });
-});
-
 describe('executeProphecy', () => {
-  function makeProphecyState(overrides: Partial<CriticalState> = {}): CriticalState {
+  function makeState(overrides: Partial<CriticalState> = {}): CriticalState {
     return {
       deck: [
         'strike',
@@ -376,37 +82,9 @@ describe('executeProphecy', () => {
     };
   }
 
-  function makeAdvancingHelpers(): EngineHelpers {
-    return {
-      addLog: (state: CriticalState, entry: GameLogEntry) => {
-        state.logs.push(entry as CriticalState['logs'][number]);
-      },
-      createLogEntry: (
-        _type: string,
-        message: string,
-        options?: Record<string, unknown>,
-      ): GameLogEntry => ({
-        id: 'test-id',
-        type: 'action',
-        message,
-        createdAt: new Date().toISOString(),
-        ...(options ?? {}),
-      }),
-      advanceTurn: (state: CriticalState) => {
-        state.currentTurnIndex =
-          (state.currentTurnIndex + state.playDirection + state.playerOrder.length) %
-          state.playerOrder.length;
-        state.pendingDraws = 1;
-      },
-      shuffleArray: () => {},
-      findPlayer: (state: CriticalState, playerId: string) =>
-        state.players.find((p) => p.playerId === playerId),
-    };
-  }
-
   it('sends private log with top 5 cards', () => {
-    const state = makeProphecyState();
-    const helpers = makeAdvancingHelpers();
+    const state = makeState();
+    const helpers = makeHelpers();
 
     executeProphecy(state, 'playerA', helpers);
 
@@ -422,8 +100,8 @@ describe('executeProphecy', () => {
   });
 
   it('sets pendingProphecy state with playerId and top5', () => {
-    const state = makeProphecyState();
-    const helpers = makeAdvancingHelpers();
+    const state = makeState();
+    const helpers = makeHelpers();
 
     executeProphecy(state, 'playerA', helpers);
 
@@ -440,8 +118,8 @@ describe('executeProphecy', () => {
   });
 
   it('fails if deck is empty', () => {
-    const state = makeProphecyState({ deck: [] });
-    const helpers = makeAdvancingHelpers();
+    const state = makeState({ deck: [] });
+    const helpers = makeHelpers();
 
     const result = executeProphecy(state, 'playerA', helpers);
 
@@ -450,10 +128,10 @@ describe('executeProphecy', () => {
   });
 
   it('peeks at fewer than 5 if deck has fewer cards', () => {
-    const state = makeProphecyState({
+    const state = makeState({
       deck: ['strike', 'evade', 'reorder'] as CriticalCard[],
     });
-    const helpers = makeAdvancingHelpers();
+    const helpers = makeHelpers();
 
     executeProphecy(state, 'playerA', helpers);
 
@@ -462,7 +140,7 @@ describe('executeProphecy', () => {
 });
 
 describe('executeCommitProphecy', () => {
-  function makeCommitState(overrides: Partial<CriticalState> = {}): CriticalState {
+  function makeState(overrides: Partial<CriticalState> = {}): CriticalState {
     const top5: CriticalCard[] = ['strike', 'evade', 'reorder', 'insight', 'cancel'];
     return {
       deck: ['strike', 'evade', 'reorder', 'insight', 'cancel', 'trade', 'neutralizer'] as CriticalCard[],
@@ -488,37 +166,9 @@ describe('executeCommitProphecy', () => {
     };
   }
 
-  function makeAdvancingHelpers(): EngineHelpers {
-    return {
-      addLog: (state: CriticalState, entry: GameLogEntry) => {
-        state.logs.push(entry as CriticalState['logs'][number]);
-      },
-      createLogEntry: (
-        _type: string,
-        message: string,
-        options?: Record<string, unknown>,
-      ): GameLogEntry => ({
-        id: 'test-id',
-        type: 'action',
-        message,
-        createdAt: new Date().toISOString(),
-        ...(options ?? {}),
-      }),
-      advanceTurn: (state: CriticalState) => {
-        state.currentTurnIndex =
-          (state.currentTurnIndex + state.playDirection + state.playerOrder.length) %
-          state.playerOrder.length;
-        state.pendingDraws = 1;
-      },
-      shuffleArray: () => {},
-      findPlayer: (state: CriticalState, playerId: string) =>
-        state.players.find((p) => p.playerId === playerId),
-    };
-  }
-
   it('reorders top 2 cards in chosen order', () => {
-    const state = makeCommitState();
-    const helpers = makeAdvancingHelpers();
+    const state = makeState();
+    const helpers = makeHelpers();
 
     executeCommitProphecy(state, 'playerA', ['evade', 'strike'], helpers);
 
@@ -527,8 +177,8 @@ describe('executeCommitProphecy', () => {
   });
 
   it('clears pendingProphecy after commit', () => {
-    const state = makeCommitState();
-    const helpers = makeAdvancingHelpers();
+    const state = makeState();
+    const helpers = makeHelpers();
 
     executeCommitProphecy(state, 'playerA', ['evade', 'strike'], helpers);
 
@@ -536,8 +186,8 @@ describe('executeCommitProphecy', () => {
   });
 
   it('advances turn after commit', () => {
-    const state = makeCommitState();
-    const helpers = makeAdvancingHelpers();
+    const state = makeState();
+    const helpers = makeHelpers();
 
     executeCommitProphecy(state, 'playerA', ['evade', 'strike'], helpers);
 
@@ -545,8 +195,8 @@ describe('executeCommitProphecy', () => {
   });
 
   it('fails if no pendingProphecy', () => {
-    const state = makeCommitState({ pendingProphecy: undefined });
-    const helpers = makeAdvancingHelpers();
+    const state = makeState({ pendingProphecy: undefined });
+    const helpers = makeHelpers();
 
     const result = executeCommitProphecy(state, 'playerA', ['evade', 'strike'], helpers);
 
@@ -555,13 +205,13 @@ describe('executeCommitProphecy', () => {
   });
 
   it('fails if pendingProphecy belongs to different player', () => {
-    const state = makeCommitState({
+    const state = makeState({
       pendingProphecy: {
         playerId: 'playerB',
         top5: ['strike', 'evade', 'reorder', 'insight', 'cancel'],
       },
     });
-    const helpers = makeAdvancingHelpers();
+    const helpers = makeHelpers();
 
     const result = executeCommitProphecy(state, 'playerA', ['evade', 'strike'], helpers);
 
@@ -569,8 +219,8 @@ describe('executeCommitProphecy', () => {
   });
 
   it('fails if reorderedTop2 does not contain cards from top5', () => {
-    const state = makeCommitState();
-    const helpers = makeAdvancingHelpers();
+    const state = makeState();
+    const helpers = makeHelpers();
 
     const result = executeCommitProphecy(
       state,
@@ -584,11 +234,27 @@ describe('executeCommitProphecy', () => {
   });
 
   it('fails if reorderedTop2 does not have exactly 2 cards', () => {
-    const state = makeCommitState();
-    const helpers = makeAdvancingHelpers();
+    const state = makeState();
+    const helpers = makeHelpers();
 
     const result = executeCommitProphecy(state, 'playerA', ['strike'] as CriticalCard[], helpers);
 
     expect(result.success).toBe(false);
+  });
+
+  it('rejects duplicate cards that bypass validation', () => {
+    // top5 has only one 'strike', so ['strike', 'strike'] should be invalid
+    const state = makeState();
+    const helpers = makeHelpers();
+
+    const result = executeCommitProphecy(
+      state,
+      'playerA',
+      ['strike', 'strike'] as CriticalCard[],
+      helpers,
+    );
+
+    expect(result.success).toBe(false);
+    expect(result.error).toBeDefined();
   });
 });
